Type:fix
Description:Bug ID:18155
     in App settings view,clicked the recommended button,then reboot App,appears some checkbox options not checked.

RootCause:Null

Solution:maybe the dsettings have problem when exit the setting dialog.
I new a member QMap to store the setting keys and value.
write to the disk when exit the setting dialog.
read from the disk when open the setting dialog.
keep same data with the DSettings's data in other conditions.