Type:fix
Description:

    the bug is : when extract the rar,7z,zip encryption files, the User input the wrong password also produce the files in the destination path.

RootCause:Null

Solution:

    We need to analyse the information about the compressed files before we execute extract operation for  the rar,7z,zip compressed files,
if User need to input the password, then the program need to analyse the important informations of the lines in the compressed files.
the information includes the format,the password which user input whether is correct or not.
    At the last ,according to the information above.The program can popup tips when the password is wrong.it also can execute extract operation when
the password is right. If the compressed files don't need password,the program just execute the extract operation.